+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 * Make popup modal bar appear in correct position on portrait 7" displays
 * Show HDOP (horizontal position precision) in the gps display
 * Only enable arm menu item if we have a heartbeat from the vehicle
+* Show the bluetooth device connect/disconnect button in the new SimpleView
 
 # 1.7.04
 * Announce Arm/Disarm via voice